Fixed a few bugs, the program should now be in a working state.

+++ src/index.js
@@ -1,8 +1,8 @@
 import * as DOM from './dom.js';
 let domObj = DOM.domBox();
-let projHolder;
-JSON.parse(localStorage.getItem('projHolder')).length >= 1 ? projHolder = JSON.parse(localStorage.getItem('projHolder')) : projHolder = [['Default', {title:'Press "Display" To See Details', description:'The "+" next to "Project List" adds a project, the "+" next to "Task List" adds a task. If you want to delete a project or a task, press the "x" or "delete" buttons.', dueDate:'N/A', priority:'High'}]];
-let lastClicked = projHolder[0][0] || 'default';
+let initialValue = [['default', {title:'Press "Display" To See Details', description:'The "+" next to "Project List" adds a project, the "+" next to "Task List" adds a task. If you want to delete a project or a task, press the "x" or "delete" buttons.', dueDate:'N/A', priority:'High'}]];
+let projHolder = initialValue;
+let lastClicked = 'default';
 
 const funcBox = () => {
     let taskPromptBtn = document.querySelector('.show-task-prompt');
@@ -15,7 +15,8 @@ const funcBox = () => {
     
     // pushes a  list item into the project with a given name
     const createProject = (projectName) => {
-        projHolder.push([projectName]);
+        projHolder.push([projectName.toLowerCase()]);
+        makeProjectInStorage(projectName);
         showProjects();
     }
     
@@ -24,9 +25,37 @@ const funcBox = () => {
         for(let i = 0; i < projHolder.length; i++) {
             if(projHolder[i].indexOf(location) >= 0) {
                 projHolder[i].push({ title, description, dueDate, priority });
-                return;
+                makeTaskInStorage(location, title, description, dueDate, priority);
             } else if (i === projHolder.length - 1 && projHolder[i].indexOf(location) === -1) {                
                 projHolder[0].push({ title, description, dueDate, priority });
+                makeTaskInStorage(location, title, description, dueDate, priority);
+            }
+        }
+    }
+
+    const checkLocalStorage = () => {
+        if(!(localStorage.getItem('projHolder'))) {
+            localStorage.setItem('projHolder', JSON.stringify(initialValue));
+        }
+    }
+
+    const makeProjectInStorage = (name) => {
+        checkLocalStorage();
+        let storageProjHolder = JSON.parse(localStorage.getItem('projHolder'));
+        storageProjHolder.push([name.toLowerCase()]);
+        localStorage.setItem('projHolder', JSON.stringify(storageProjHolder));
+    }
+
+    const makeTaskInStorage = (location='default', title, description, dueDate, priority) => {
+        checkLocalStorage();
+        let storageProjHolder = JSON.parse(localStorage.getItem('projHolder'));
+        for(let i = 0; i < storageProjHolder.length; i++) {
+            if(storageProjHolder[i].indexOf(location) >= 0) {
+                storageProjHolder[i].push({ title, description, dueDate, priority });
+                localStorage.setItem('projHolder', JSON.stringify(storageProjHolder));
+            } else if (i === storageProjHolder.length - 1 && storageProjHolder[i].indexOf(location) === -1) {                
+                storageProjHolder[0].push({ title, description, dueDate, priority });
+                localStorage.setItem('projHolder', JSON.stringify(storageProjHolder));
             }
         }
     }
@@ -41,17 +70,37 @@ const funcBox = () => {
     // clears the task list each time it's run so that we don't have tasks from different projects contaminating the ui
     // afterwards iterates over projHolder and compares the first value to the textContent of the button
     // if it matches it runs appendToTaskList to append the values to the innerHTML of the ul
-    const showAttributes = (e='') => {
+    const showAttributes = (e) => {
         domObj.clearTaskList();
-        if(e!=='') {
-            lastClicked = e.target.childNodes[0].textContent;
-            for(let i = 0; i < projListItems.length; i++) {
-                projListItems[i].classList.remove('current');
-                projListItems[i].childNodes[1].classList.remove('button-primary');
+        setCurrentBtn(e);
+        lastClicked = e.target.childNodes[0].textContent;
+        
+        for(let i = 0; i < projHolder.length; i++) {
+            if(projHolder[i][0] === lastClicked) {
+                for(let j = 1; j < projHolder[i].length; j++) {
+                    domObj.appendToTaskList(projHolder[i][j].title, projHolder[i][j].description, projHolder[i][j].dueDate, projHolder[i][j].priority )
+                }
             }
+        }
+        removeFirstDelete();
+        deleteTaskEvent();
+        domObj.displayDetailsEvent();
+    }
+
+    const setCurrentBtn = (e) => {
+        for(let i = 0; i < projListItems.length; i++) {
+            projListItems[i].classList.remove('current');
+            projListItems[i].childNodes[1].classList.remove('button-primary');
+        }
+        if(e.target.childNodes[1]) {
             e.target.classList.add('current');
             e.target.childNodes[1].classList.add('button-primary');
         }
+    }
+
+    const loadAttributes = () => {
+        domObj.clearTaskList();
+        lastClicked = 'default';
         for(let i = 0; i < projHolder.length; i++) {
             if(projHolder[i][0] === lastClicked) {
                 for(let j = 1; j < projHolder[i].length; j++) {
@@ -71,6 +120,7 @@ const funcBox = () => {
         showAttrEvent();
         setFirstAsCurrent();
         deleteProjectEvent();
+        removeFirstDelete();
     }
 
     // asks the user for the title name and creates a project based on the given name
@@ -80,19 +130,27 @@ const funcBox = () => {
         newProjBtn.addEventListener('click', (e) => {
             e.preventDefault();
             let title = document.getElementById('project-title').value;
-            if(title.trim() !== '') {
+            if(title.trim() !== '' && checkIfExists(title, projHolder) !== true) {
                 createProject(title)
                 domObj.hideWindow(projectPromptWindow);
                 deleteProjectEvent();
-                saveToLocalStorage();
+            } else if(checkIfExists(title, projHolder)) {
+                domObj.showError('Project already exists.', projectPromptWindow, projectForm);
             } else {
-                domObj.showError('Project name cannot be blank.', projectPromptWindow, projectForm);
+                domObj.showError('Please enter a value for project name.', projectPromptWindow, projectForm);
             }
         });
     }
 
+    const checkIfExists = (value, location) => {
+        for(let i = 0; i < location.length; i++) {
+            if(location[i].includes(value.trim().toLowerCase())) {
+                return true;
+            }
+        }
+    }
+
     // asks for a series of prompts and then calls the createTask function based on the answers
-    // the location is set to the lastClicked element which is set every time showAttributes is called
     const createTaskEvent = () => {
         let newTaskBtn = document.querySelector('.new-task');
         
@@ -103,19 +161,20 @@ const funcBox = () => {
             let description = document.getElementById('desc');
             let dueDate = document.getElementById('date');
             let priority = document.getElementById('priority');
-            if(description.value.trim() === '' && title.value.trim() === '') {
-                domObj.showError('Please fill out the title and description inputs.', taskPromptWindow, taskForm);
+            if(description.value.trim() === '' && title.value.trim() === '' && !dueDate.value) {
+                domObj.showError('Please fill out the description, title and include a valid date.', taskPromptWindow, taskForm);
             } else if(description.value.trim() === '') {
                 domObj.showError('Please fill out the description input.', taskPromptWindow, taskForm);
             } else if(title.value.trim() === '') {
                 domObj.showError('Please fill out the title input.', taskPromptWindow, taskForm);
+            } else if(!dueDate.value) {
+                domObj.showError('Please enter a valid date.', taskPromptWindow, taskForm);
             } else {
                 createTask(lastClicked, title.value, description.value, dueDate.value, priority.value );
                 domObj.appendToTaskList(title.value, description.value, dueDate.value, priority.value )
                 domObj.displayDetailsEvent();
                 domObj.hideWindow(taskPromptWindow);
                 deleteTaskEvent();
-                saveToLocalStorage();            
             }
             // after we've submitted our task, the window closes
         });
@@ -128,8 +187,8 @@ const funcBox = () => {
         for(let i = 0; i < deleteTaskBtn.length; i++) {
             deleteTaskBtn[i].addEventListener('click', (e) => { 
                 deleteTaskFromUI(e);
-                deleteTaskFromDatabase(e);                
-                saveToLocalStorage();
+                deleteTaskFromDatabase(e);     
+                deleteTaskFromLS(e);   
             });
         }
     }
@@ -154,13 +213,33 @@ const funcBox = () => {
         }
     }
 
-    const deleteProjectEvent = () => {
+    const deleteTaskFromLS = (event) => {
+        checkLocalStorage();
+        let storageProjHolder = JSON.parse(localStorage.getItem('projHolder'));
+        for(let i = 0; i < storageProjHolder.length; i++) {
+            // only iterate over the values if we're in the current project
+            if(storageProjHolder[i][0].includes(lastClicked)) {                
+                for(let j = 0; j < storageProjHolder[i].length; j++) {
+                    // compare the title value of the current iteration to the title in the DOM
+                    if(storageProjHolder[i][j].title === event.target.parentElement.parentElement.childNodes[1].textContent) {
+                        storageProjHolder[i].splice(j, 1);
+                    }
+                }
+            }
+        }
+        localStorage.setItem('projHolder', JSON.stringify(storageProjHolder));
+    }
+
+    const deleteProjectEvent = (event) => {
         let deleteProjectBtn = document.getElementsByClassName('delete-project');
         for(let i = 0; i < deleteProjectBtn.length; i++) {
             deleteProjectBtn[i].addEventListener('click', (e) => {
                 deleteProjectFromUI(e);
-                deleteProjectFromDatabase(e);                
-                saveToLocalStorage();
+                deleteProjectFromDatabase(e);  
+                deleteProjectFromLS(e);
+                setFirstAsCurrent();
+                loadAttributes();
+                e.stopPropagation();
             });
         }
     }
@@ -173,31 +252,41 @@ const funcBox = () => {
         for(let i = 0; i < projHolder.length; i++) {
             if(projHolder[i][0].includes(e.target.parentElement.childNodes[0].textContent)) {
                 projHolder.splice(i, 1);
-                lastClicked = 'default';
             }
         }
     }
 
-    const saveToLocalStorage = () => {
-        localStorage.setItem('projHolder', JSON.stringify(projHolder));
+    const deleteProjectFromLS = (e) => {
+        checkLocalStorage();
+        let storageProjHolder = JSON.parse(localStorage.getItem('projHolder'));
+        for(let i = 0; i < storageProjHolder.length; i++) {
+            if(storageProjHolder[i][0].includes(e.target.parentElement.childNodes[0].textContent)) {
+                storageProjHolder.splice(i, 1);
+            }
+        }
+        localStorage.setItem('projHolder', JSON.stringify(storageProjHolder));
     }
 
     const setFirstAsCurrent = () => {
         projListItems[0].classList.add('current');
-        projListItems[0].childNodes[1].classList.add('button-primary'); 
         lastClicked = projHolder[0][0];
     }
 
     const run = () => {
+        checkLocalStorage();
+        projHolder = JSON.parse(localStorage.getItem('projHolder'));
+        domObj.displayDetailsEvent();
         createProjectEvent();
         createTaskEvent();
+        deleteProjectEvent();
+        deleteTaskEvent();
         taskPromptBtn.addEventListener('click', (e) => {domObj.taskToggler(taskPromptWindow)});
         newProjectBtn.addEventListener('click', (e) => {domObj.projectToggler(projectPromptWindow)});
         showProjects();
-        showAttributes();
         if(projHolder !== undefined && projHolder.length >= 1) {
             setFirstAsCurrent();
         }
+        loadAttributes();
         removeFirstDelete();
     }
